Implemented mutation function, and generation count

+++ BallShooter.cs
@@ -69,7 +69,7 @@ namespace PeggleAI
         }
 
         // This override will shoot the ball at a given angle, given in degrees
-        public void shoot(int angle)
+        public void shoot(double angle)
         {
             float radAngle = (float)(Math.PI * angle / 180.0);
             this.ball = new Ball(0, 3.6f, radAngle, level.world);
